Ensure other attributes are ignored
If an application includes a link tag with attributes other than
`rel="stylesheet"` and `href="something"` the regex to find the link
fails.  This allows any attribute combo to come before href and still
match.

This was particularly problematic when dealing with links that include
`type="text/css"` for example.